Get rid of cyclic imports
