+++ src/api/worker/search/WhitelabelChildIndexer.ts
@@ -1,20 +1,16 @@
 import { FULL_INDEXED_TIMESTAMP, NOTHING_INDEXED_TIMESTAMP, OperationType } from "../../common/TutanotaConstants"
 import { NotFoundError } from "../../common/error/RestError"
-import type { WhitelabelChild } from "../../entities/sys/TypeRefs.js"
-import { WhitelabelChildTypeRef } from "../../entities/sys/TypeRefs.js"
-import { neverNull, noOp } from "@tutao/tutanota-utils"
+import type { EntityUpdate, User, WhitelabelChild } from "../../entities/sys/TypeRefs.js"
+import { CustomerTypeRef, WhitelabelChildTypeRef } from "../../entities/sys/TypeRefs.js"
+import { neverNull, noOp, ofClass, promiseMap } from "@tutao/tutanota-utils"
 import type { Db, GroupData, IndexUpdate, SearchIndexEntry } from "./SearchTypes"
 import { _createNewIndexUpdate, typeRefToTypeInfo, userIsGlobalAdmin } from "./IndexUtils"
-import { CustomerTypeRef } from "../../entities/sys/TypeRefs.js"
-import { GroupDataOS } from "./Indexer"
 import { IndexerCore } from "./IndexerCore"
 import { SuggestionFacade } from "./SuggestionFacade"
 import { tokenize } from "./Tokenizer"
-import type { EntityUpdate } from "../../entities/sys/TypeRefs.js"
-import type { User } from "../../entities/sys/TypeRefs.js"
 import { EntityClient } from "../../common/EntityClient"
-import { ofClass, promiseMap } from "@tutao/tutanota-utils"
 import { typeModels } from "../../entities/sys/TypeModels"
+import { GroupDataOS } from "./IndexTables.js"
 
 export class WhitelabelChildIndexer {
 	_core: IndexerCore
