+++ src/calendar/view/MultiDayCalendarView.ts
@@ -30,14 +30,13 @@ import { PageView } from "../../gui/base/PageView"
 import type { CalendarEvent } from "../../api/entities/tutanota/TypeRefs.js"
 import { logins } from "../../api/main/LoginController"
 import type { GroupColors } from "./CalendarView"
-import { SELECTED_DATE_INDICATOR_THICKNESS } from "./CalendarView"
 import type { EventDragHandlerCallbacks, MousePos } from "./EventDragHandler"
 import { EventDragHandler } from "./EventDragHandler"
 import { getPosAndBoundsFromMouseEvent } from "../../gui/base/GuiUtils"
 import { UserError } from "../../api/main/UserError"
 import { showUserError } from "../../misc/ErrorHandlerImpl"
 import { styles } from "../../gui/styles"
-import { renderCalendarSwitchLeftButton, renderCalendarSwitchRightButton } from "./CalendarGuiUtils"
+import { renderCalendarSwitchLeftButton, renderCalendarSwitchRightButton, SELECTED_DATE_INDICATOR_THICKNESS } from "./CalendarGuiUtils"
 import type { CalendarEventBubbleClickHandler, EventsOnDays } from "./CalendarViewModel"
 import { CalendarViewType } from "./CalendarViewModel"
 import { ContinuingCalendarEventBubble } from "./ContinuingCalendarEventBubble"
