+++ src/settings/contactform/ContactFormListView.ts
@@ -7,21 +7,18 @@ import { NotFoundError } from "../../api/common/error/RestError"
 import { size } from "../../gui/size"
 import type { SettingsView, UpdatableSettingsViewer } from "../SettingsView"
 import { LazyLoaded, neverNull, ofClass } from "@tutao/tutanota-utils"
-import { ContactFormViewer, getContactFormUrl } from "./ContactFormViewer"
+import { ContactFormViewer } from "./ContactFormViewer"
 import * as ContactFormEditor from "./ContactFormEditor"
 import type { ContactForm } from "../../api/entities/tutanota/TypeRefs.js"
-import { ContactFormTypeRef } from "../../api/entities/tutanota/TypeRefs.js"
+import { ContactFormTypeRef, CustomerContactFormGroupRootTypeRef } from "../../api/entities/tutanota/TypeRefs.js"
 import { getWhitelabelDomain } from "../../api/common/utils/Utils"
-import { CustomerTypeRef } from "../../api/entities/sys/TypeRefs.js"
 import type { CustomerInfo } from "../../api/entities/sys/TypeRefs.js"
-import { CustomerInfoTypeRef } from "../../api/entities/sys/TypeRefs.js"
+import { CustomerInfoTypeRef, CustomerTypeRef } from "../../api/entities/sys/TypeRefs.js"
 import { logins } from "../../api/main/LoginController"
-import { Dialog } from "../../gui/base/Dialog"
 import { OperationType } from "../../api/common/TutanotaConstants"
 import { Icon } from "../../gui/base/Icon"
 import { Icons } from "../../gui/base/icons/Icons"
-import { CustomerContactFormGroupRootTypeRef } from "../../api/entities/tutanota/TypeRefs.js"
-import { getAdministratedGroupIds, getDefaultContactFormLanguage } from "./ContactFormUtils"
+import { getAdministratedGroupIds, getContactFormUrl, getDefaultContactFormLanguage } from "./ContactFormUtils"
 import type { EntityUpdateData } from "../../api/main/EventController"
 import { isUpdateForTypeRef } from "../../api/main/EventController"
 import { Button, ButtonType } from "../../gui/base/Button.js"
