+++ src/subscription/SubscriptionViewer.ts
@@ -16,7 +16,7 @@ import { assertNotNull, downcast, incrementDate, neverNull, noOp, ofClass, promi
 import { logins } from "../api/main/LoginController"
 import { lang, TranslationKey } from "../misc/LanguageViewModel"
 import { Icons } from "../gui/base/icons/Icons"
-import { asPaymentInterval, formatPrice, formatPriceDataWithInfo, getCurrentCount, PriceAndConfigProvider, PaymentInterval } from "./PriceUtils"
+import { asPaymentInterval, formatPrice, formatPriceDataWithInfo, PaymentInterval, PriceAndConfigProvider } from "./PriceUtils"
 import { formatDate, formatNameAndAddress, formatStorageSize } from "../misc/Formatter"
 import { getByAbbreviation } from "../api/common/CountryList"
 import * as AddUserDialog from "../settings/AddUserDialog"
@@ -33,7 +33,7 @@ import * as SwitchToBusinessInvoiceDataDialog from "./SwitchToBusinessInvoiceDat
 import { NotFoundError } from "../api/common/error/RestError"
 import type { EntityUpdateData } from "../api/main/EventController"
 import { isUpdateForTypeRef } from "../api/main/EventController"
-import { getTotalAliases, getTotalStorageCapacity, isBusinessFeatureActive, isSharingActive, isWhitelabelActive } from "./SubscriptionUtils"
+import { getCurrentCount, getTotalAliases, getTotalStorageCapacity, isBusinessFeatureActive, isSharingActive, isWhitelabelActive } from "./SubscriptionUtils"
 import { Button, ButtonType } from "../gui/base/Button.js"
 import { TextField } from "../gui/base/TextField.js"
 import { Dialog, DialogType } from "../gui/base/Dialog"
