+++ src/subscription/UpgradeSubscriptionWizard.ts
@@ -22,7 +22,7 @@ import { StorageBehavior } from "../misc/UsageTestModel"
 import { UpgradePriceService } from "../api/entities/sys/Services.js"
 import { FeatureListProvider, SelectedSubscriptionOptions, SubscriptionType } from "./FeatureListProvider"
 import { UpgradeType } from "./SubscriptionUtils"
-import { asPaymentInterval, PriceAndConfigProvider, PaymentInterval } from "./PriceUtils"
+import { asPaymentInterval, PaymentInterval, PriceAndConfigProvider } from "./PriceUtils"
 
 assertMainOrNode()
 export type SubscriptionParameters = {
@@ -31,13 +31,6 @@ export type SubscriptionParameters = {
 	interval: string // typed as string because m.parseQueryString returns an object with strings
 }
 
-/** Subscription type passed from the website */
-export const SubscriptionTypeParameter = Object.freeze({
-	FREE: "free",
-	PREMIUM: "premium",
-	TEAMS: "teams",
-	PRO: "pro",
-})
 export type NewAccountData = {
 	mailAddress: string
 	recoverCode: Hex
