+++ src/templates/view/TemplateConstants.ts
@@ -1,0 +1,4 @@
+export const TEMPLATE_POPUP_HEIGHT = 340
+export const TEMPLATE_POPUP_TWO_COLUMN_MIN_WIDTH = 600
+export const TEMPLATE_LIST_ENTRY_HEIGHT = 47
+export const TEMPLATE_LIST_ENTRY_WIDTH = 354
