+++ src/templates/view/TemplatePopupResultRow.ts
@@ -1,8 +1,8 @@
 import m, { Children, Component, Vnode } from "mithril"
-import { TEMPLATE_LIST_ENTRY_HEIGHT } from "./TemplatePopup"
 import { px } from "../../gui/size"
 import type { EmailTemplate } from "../../api/entities/tutanota/TypeRefs.js"
 import { TEMPLATE_SHORTCUT_PREFIX } from "../model/TemplatePopupModel"
+import { TEMPLATE_LIST_ENTRY_HEIGHT } from "./TemplateConstants.js"
 
 export type TemplateResultRowAttrs = {
 	template: EmailTemplate
